Move MongoJsonDeserializationSchema into test case
